don't do that hacky nonsense with less
js.lesscss will try to compile your files if it finds LESSC in the
environment, but will serve the minified version if not. That means
I may have to re-instate this hack in the debug run script, but at
least for production it should be left out in order to serve the
minified css.