tighten shadows and transitions for a cleaner look
I found interesting information about scrolling smoothness and
the expense of painting blurred box shadows:
http://nerds.airbnb.com/box-shadows-are-expensive-to-paint

As a result, I shrunk the smallshadow mixin to be a fixed 3px blur
radius instead of 1em.

I stripped the box shadow off the heatmap entirely since it could
decrease scrolling performance on the source page for little visual
gain.

Finally, I took out the shadow transitions since it was creating
visual artifacts in webkit on the thread expanders and is probably
hell of unneccessarily expensive.

I believe this fixes #107.