brand spelling consistency, OSS messaging
This welcome page is not the place for details about signing up
for our alpha. The application is designed to be extensible and
customizable. We should deploy a separate project that depends upon
this one, which contains a custom authentication connector for our
pre-authenticated users and a custom home page template.

See, for example, annotateit vs annotator.

This should really be a first-run page for developers who are starting
up their own instance.