allow local consumer to have a provided secret
A half-baked idea, but perhaps there is a use case where the app
needs to verify a token but does not have access to the persistent
storage for looking up Consumer credentials. It may even make more
sense as a default setup than auto-generating a secret and storing
it in the database.