manage an XSRF-TOKEN cookie
AngularJS expects a JavaScript-readable cookie named 'XSRF-TOKEN'
to contain the cross site request forgery token. Since Pyramid
stores this in the session, introduce a custom session factory
which uses `pyramid_beaker` but transparently manages this extra
cookie so that it always matches the value in the session. Add a
subscriber so that POST requests are automatically updated with
the value Angular sends in the request headers.

This makes passing the token in a hidden field of the forms
obsolete when AngularJS is present.