move fuzzyTime and userName into template
The annotation directive is left to worry only about the text.
Later, it can worry about permissions and other mutable properties.

Moving fuzzyTime into the template has the added benefit that it
will be re-rendered on root $apply so the times can be easily kept
up to date in the UI.