pull dynamic bucket behavior out of viewer
We can start to think of the bucket view as the base panel
and think about moving the viewer and the editor into the
sliding panels. The idea is that then the bucket view is behind
the editor as well as the detail view and the current viewer
becomes detail-only while the bucket is always underneath.

Also, the heatmap is back to adding itself to the element so that
it should be possible to move the heatmap out of the iframe
at some point soon, making it possible to absolute position it
with the full height of the page. This will also reduce postMessage
chatter when scrolling/updating.