improve sleepiness of minimized state
Reduce the number of places where we do tricky things with the
location and state to avoid opening the sidebar accidentally and
restore the Annotator default of show(Editor|Viewer) causing the
UI to be shown (panel restored to expanded position).

As a side bonus, do less processing when the annotator is minimized
by using that state correctly.