break apart the annotation directive
Establish privacy and markdown directives. The reason is that
ngModelController uses strict equality when determining that the
parsed view value is different from the model value. Since Annotator
relies on the object identity of annotations to remain stable, it
was impractical to trigger a real view change on an annotation.

With this change, constituent values of the annotations are managed
in their own model controllers to better encapsulate the various
UI controls that make up an annotation.

In addition, some attention was put into cleaning up the templates
and clarifying the interactions with annotator events and the digest
cycle. The result is much better, I think.