standard header typography and annotation styling
We were using header tags for the annotation user name and time for
which we mangled the typography to fit this use. This restores the
suggested UA styles for header elements and reworks the annotation
user line to fit the way it was before.

In addition, the time is moved up to the top line.