avoid loading two annotators
This is a little tricky. The reason the checks are in multiple places
is that content scripts don't share the window object of the host page.
Therefore, when running in the Chrome extension, we should make sure
to set the annotator object on *both* window objects. The one in the
content script context will have a null annoator which tests as not
undefined (preventing double loading the extension), but a test still
needs to be performed in the bootstrap script in case the page already
has an annotator.

In a non-extension conflict, window.annotator goes from undefined to
null to an object and, finally, an Annotator.