more flexible head slot
Using metal:fill-slot in combination with tal:omit-tag makes it easy
to provide a slot to add any number of child elements to a node. In
this case, it's used to make the <head> tag in the base template
extensible by the layout templates.

Also, make the embed deferred on the home site.