add base_url and service_url to app
base_url communicates the base URL of the application to the page
using the `<base>` tag. With this change the extension can render
the app template into the extension bundle and ensure that the
extension still knows the base URL for the live application.

service_url is mainly added to get more logic out of the template
and let the app pass it in.