no reason to use threading and Queue for streaming
pyramid_sockjs uses gevent for greening the socket operations and
when running on a fully green WSGI container like gunicorn+gevent
each of these sends is an implicit context switch. Therefore, it's
not problematic to loop over all connections and send and there is
no need for a separate background thread for this.

If desired, the appropriate place for this type of thing would be
in the domain metadata fetching in order to consolidate requests
for the same URI. I suspect some of that BeautifulSoup stuff can
actually disappear completely now that @edsu's Document plugin is
storing document metadata along with annotations.