Introduce an AnnotatorStoreEvent into registry
With this change the streamer and api.store modules are kept separate
by having them interact through registry events. This preserves the
ability to run the application without an embedded store (this may
not yet be possible, but it's close, and eventually will include an
implementation that delegates via HTTP to an external service). With
this change the annotator-store no longer needs all the 'before_' and
'after_' events for each store route. Instead, we spy on the endpoint
in a generic 'after_request' to determine which action happened and
broadcast the result with the new registry event.

An added benefit of the registry event is that we can add a pub-sub
component on the server as a separate module that is optionally
imported in multi-host deployments without polluting any of this code.