Break out bookmarklet template and tweak it.
I changed the message to not be specific to Hypothes.is and to not
reference the development docs. We really want people contacting
whomever is in charge of the service they're using, not necessarily
the developers, because whether it's usable on SSL is a deployment
issue, not a code issue.