CSS clean up: replace gray-tags with descendant selectors.
Some bg: According to our CSS, body and .noise are gray,
and .paper is white. Therefore, on .noise, we want white tags,
but on .paper, we want gray tags.

To make things more interesting, .noise and .paper like to
be nested in each other, alternating the bg color ad nauseam.
Therefore I could come up with a proper descendant selector
expression (not even using the :not pseode-selector), so
I have just hard-wired the first 6 levels of recursion.

Feel free to improve.

   * * *

Anyway, the templates are now clear; no special marking classes
are necessary for configuring tags.