Use the server_name for local id authority
The dev server now explicitly listens on localhost. That means that
if you wish to expose this publicly you will need to set up a public-
facing proxy server, such as nginx, or create/modify a .ini file.

The Procfile still uses 0.0.0.0, which means that pyramid will fall
back to reading the HTTP Host header for the server name. In Procfile
deployments like heroku and dokku it is assumed that this header is
trusted and fixed in upstream reverse proxies.

Fixes #497.