Added framework for Social View support, with single player demo
Now the Hypothes.is class has a field called 'socialView',
which specifies the social view filters that should be used.

This is considered when loading annotations on init, and also on
later re-loads, because of login/logout.

When the filter is changed, annotations are automatically reloaded.

(And there were also some cosmetic fixes.)