Use a directory-local pip download cache
It's not very nice to drop a bunch of tarballs, potentially not small,
into a hidden directory in a user's home folder. Instead, let's honor
the environment variable, PIP_DOWNLOAD_CACHE, if it exists, but if it
does not we shall use a local directory. At least this way, if the
user decides to delete the repository, she will have removed all the
downloaded packages as well and we will not have left cruft on her
machine.