Consolidate app and views modules
This distinction was never useful. Initially, I had thought that there
would be a site and an app but now it's clear that the site is an app
and this distinction is unnecessary. The file is not so long that it
shouldn't just be one and it keeps all the controllers together which
might make it easier for newcomers to find things.