Fix closing the searchbar in situations when the search was never executed
When the user runs a search, the sidebar switches to search view.
When killing the searchbar (with the x) afterwards, it switches back to viewer.
Leaving search mode triggers a route change watch,
which executes the necessary cleanup.

However, if we haven't yet started searching, only opened the searchbar,
and then kill it immediately (without running the search first), then
we are still in viewer mod, so no route change will happen.
Therefore, the cleanup is not triggered automatically.

So in these cases, we will have to trigger in manually.

Fixes #673.
Fixes #675.