Heatmap: introducing gapSize parameter.
The clustering algorithm uses this gapSize parameter (it was a constant before) to determine if an annotation is far enough or not from the current bucket position to merge the annotation into this bucket.
As a constant it was set to 180, but now it has a default value of 60 to make the clustering less agressive.