Bugfix: self.filter was initialized as a dict by default (instead of None)
An old error was lingering in the streamer.py The filter class variable in the StreamerSession class was initialized as a dict in the on_open() and only set to its correct value in the on_message(). If a notification happened between these two calls the after_action() call failed, because dict has no match() method.
This has been fixed.