Improve past_data load
After profiling it occured that the slowest part of the past_data preparing (~90% of the total time was wasted here) was the step which checked for the every favicon whether they really exist or not.
This is unnecessary because if they do not exist we can hide them in client side hooking to the onerror property of the img tag and favicon is no longer shown in the new stream design so only the standalone page shows them, and now it shows them correctly without backend checking.

Fixes #707