Track annotation listings by annotation id
We're still unintentionally sending internal hash keys used by angular
in our annotation payloads. I believe this is the source of #575. On
stream pages, I observed the duplicate items message when updates were
pushed and this change fixes it. Now, object identity in the repeater
is tracked by the annotation id.

Closes #575