Notify SearchController about reloaded annotations.
After login/logout if there were search results, they were not displayed again. Now a new event is broadcasted to which the SearchController listens to and calls the refresh() again.
Currently it is done with a timer. The most elegant solution would be having an event about loading and anchoring all annotations and after that we could call this event properly.

Fixes #959