Remove AppFactory in favor of RootFactory
This change sets up two changes. First, it moves to eliminate the
rather useless distinction between site and app. The app is all there
is. As a temporary hack, /app still loops back to the root resource
and the index route still overrides for /. As a result, no
functionality changes, but the index page can be easily dropped in the
future. Second, it removes the horus dependency for h.resources, which
may or may not prove to be useful.

In the process, a bug was uncovered wherein zope.interface raises
exceptions when interrogating interface support for a class which does
not implement any interfaces or an instance which provides none.
Whether the `InnerResource.__getitem__` construction and the pattern
in h.resources#includeme for creating routes is useful is still up in
the air, but at least it's more correct now.

Finally, our new RootFactory adds an ACL entry to give 'account'
permission to the recently-added principal `h.security.Authorizations`.