viewer.html edited to use thread instead of reply_list closes #1017
This may not be optimal code, as I added a repliesList function to directives
that returns messages from the thread.children array.  The other
solution I could think of would be less elegant, but also less
time-complex.  This would be to simply set the variable annotation
to hold the message property (e.g. {{annotation = annotation.message}}).