move jquery.ui.tooltip in dependent bundle
I tried to keep the core requirements on top, and the additional
modules in a separate bundle. Arbitrary, I know, but let's go with
it. It's less error prone than keeping them all in dependency order
because it's tempting to people who are obsessive like me to come
and try to alphabetize them and then break things.