IStreamResource and dict-inheriting BaseResource
Move toward unifying the stream with the rest of the app, and make
changes toward simplifying all Resources, by making the BaseResource
a subclass of dict and adding an IStreamResource interface.

With this change, it becomes possible to store arbitrary data in
every resource, which will help to clean up the currently ugly division
of template environment and model keys in the AnnotationController
view.