Kill sliding-panels and summary hover shadow
The sliding panels are totally unused. They can be brought back but
right now it's unnecessary structure. The most reasonable place to
relocate the card-emphasis was in the stream-list. After all, cards
go in streams. But I found the stream-list class wasn't actually
placed on the list, it was placed on the list items. That seems
counterintuitive, so I raised it up in the structure.

Afterward, to keep the emphasis looking correct, I changed the paper
class to not have padding, and moved that padding to the thread
instead. Then I had to change some things about the thread, including
using it as a class in the editor (which is consistent, and makes
some sense). That meant that some viewer-only tag styles were bleeding
over, so I worked on those to make the tag styles a little nicer.

Finally, I had to fix the hover shadow on the threads now, and opted
to go for a background darkening instead of a shadow, since it works
across both replies and top level threads, and makes things less
bouncy, which I find to be a relief.

All in all, I'm happy. Probably will make someone unhappy.