Add V1 viewsort feature.
The viewsort feature allows users to better control the annotations that appear in the sidebar. The view control allows users to select either SCREEN (in which the sidebar only displays annotations that are currently visible on screen) or DOCUMENT (All annotation on the page are visible in the sidebar). The sort feature allows users to sort annotations by newest, oldest, or location (where they fall in the document). Both controls are implemented in simple tabs below the topbar.

The sort feature is implemented using Angular's very useful OrderBy attribute on ng-repeat.

The view control is implemented using updateviewer.