Fix email body and subject encoding
Originally the python mailer tried to convert the email's body and subject into ascii, which failed for same unicode values.
It has been fixed to convert them to utf8 first.

Also the pyramid mailer 0.9 had a known bug for not handling encoding correctly for the same scenario, so this had to be upgraded as well.

Fixes #1106