Add uri analyzer for our uri field
Since our 'uri' field was set to not_analyzed it was not available to make a case insensitive search for ES.
This solution adds an uri analyzer to the  already multi-field uri. This analyzer uses an uax_url_email tokenizer and a lowercase filter.

After that our streaming page will use this field when searching for an uri segment.

Fixes #1028