Use showViewSort for all the things
I think this better ensures #1135 is fixed cleanly. Before, the
controls might hide after the timeout, but the user may be on their
way to click it. Now, everything goes through showViewSort, and it
manages timeout cancellation.

Also, moved the view and sort controls out to the app.pt. Because
they're essentially global now, and we should move to kill the viewer,
editor, search distinction anyway.