Collect all quote items for Page search
Originaly visual-search automatically collects the text fields into one space separated list.
But it does not do the same for any other field type, so we have to collect our search facet's values for quotes manually.

This changes now expects that our quote search results are a list and handles it like this all the way. (Even for generating the highlightQuote attribute)
Now if multiple query facets are given it assumes an 'AND' relation between them.

A nice future task will be makes this page search configurable as like the stream facets, but that should come after viewer/page search unification