:rage: bump test timeout
Seriously confused why this is failing on sauce/travis despite working
locally and showing a sauce screencast that looks correct. Going to
take a guess that it could be timeout releated. So much rage. It's really
terrible to have to have long timeouts because we would like the tests
to also ensure things are loading reasonably quickly!!!

Let's just see if we can at least get these green, though. I'm tired of
thinking our code is totally broken.