Work around horus assumptions about csrf tokens
Horus assumes the form has a csrf_token. The validator has handled
headers or form parameters since 4815084 by using the built-in csrf
check function in Pyramid, but without including the field in the
schema with a "missing" value horus will choke during registration
when it tries to delete the value before taking the rest of the
deserialized values as the attributes of the new user.