Remove Beaker sessions and h.session
I had this crazy hack here where I was hijacking the beaker session
machinery to set the XSRF-TOKEN cookie for Angular. With Pyramid 1.5,
the response callbacks are run before NewResponse subscribers so we
can greatly simplify this. Also, I was thinking at the time that it
was important to set the XSRF-TOKEN cookie with the exact same settings
as the session cookie (secure, domain, path, etc), but this probably
isn't necessary. The XSRF-TOKEN cookie can just be a simple session
cookie.

New deployments are free to use whichever session library is most
appropriate, such as pyramid_redis_sessions or pyramid_beaker, and
the XSRF-TOKEN for Angular should just work.

Also this was some of the grossest code in this repo so :sparkles:.

Closes #1063