Better messages in ./bootstrap
Thanks to @onceuponatimeforever for making it clear that the messages
emitted by the bootstrap script could be much better.

The new messages take care to reassure the developer that optional
dependencies are optional. Required dependencies cause the script to
error out early with status code 1 and now instruct the developer to
consult the installation documentation.

Close #1176