Reset Permissions plugin when logging in
In the past when we were switching users the previous user data has remained inside the Permissions.options.permissions object.
This way, our newly created annotations had all the previous logged in user's permissions in their permission list which wat totally wrong.
In this fix when logging in/out this options.permissions object is resetted too. (Along with the user)

Fixes #1179