User the term/terms ES filter properly for streams
Our streamer backend converts our operators into ES operators.
According to the ES spec, when the term: {field:value} formula is used
then the value must be a single item, not a list. For list the
terms: {field: value} filter should be used.

This fix does exactly that.

It could stay in our system so long because calling ES natively
is sometimes let's you get away with the term filter for lists too.
But the python elasticsearch library (opposed to PyES) is more strict.