Analyze annotation.user field
In order to be able to search (inside stream) for case insensitive
usernames without having that inner hack inside the streamer.py
the mapping of the user field should be changed.

Originally it was not analyzed not our custom user analyzer is added.
Which is basically a lowercase keyword analyzer.