Basic authorization check for notifications
From the original notification concept the authorization was missing.
This fix handles this in the most basic way: we do not send notification
for any private annotation at all.

This is not a final solution, just a workaround until this gets discussed.
The main difference between this authorization and our user based auth
that we here (in the long term) we should be able to tell, if
we're authorized to send an annotation to an e-mail address which is
not always connected to any user at all.

Until then, let's just not leak any private annotations at all.
Fixes #1182 for now