Formalize the use of the User.status flags
The reservation forms currently in use set two bits in the status
column of the user table. The low bit indicates confirmation of email
address, the next lowest newsletter opt-out.

This change formalizes this by adding properties to the User class.
It also changes the subscriptions boolean column to be the third lowest
bit of this field.

This allows us to run against an older installation without mapper
conflicts. In the future, we may need to do migrations for upgrades, but
for now this avoids one.