beforeAnnotationCreated before setupAnnotation
When always on highlights and the highlighter mode were introduced in
ddf17189fbf2d428b21fd31065a1c2d2bc6ab08b the beforeAnnotationCreated
event was not fired in the guest frame. Instead, a special inject call
was added to the bridge to send the annotation and the event fired in
the iframe first. Later, when I removed this, I added the event after
the call to setupAnnotation.

The change in 3613c7d91682a438b1d243b9573cb941b26cdc43 uses the
annotationsLoaded event to notify listeners when highlights change as a
result of realizing anchors. This presents a problem because the bridge
ignores the beforeAnnotationCreated event for annotations that have
already been tagged by the bridge, but this load event causes the new
highlight to be tagged and serialized.

Move the beforeAnnotationCreated event to before the setupAnnotation
call, which is the order of things in non-highlighter mode, so that
the event is properly propagated to the sidebar iframe.

Close #1240