Fix mappings for tags field
The tags field contained the wrong index_name: not_analyzed property
so because of this, this field was analyzed with the default analyzer
which split the search expression at hyphens.

Because we want to do case insensitive searching for this field,
the same analyzer is needed (instead of the intented not-analyzed)
as used for the user name.

Fixes #1251