Fix failing annotator login test on OSX
The SERVER_NAME environment variable on OSX defaults to 127.0.0.1 where
on other systems it's localhost. When a user is created the provider
defaults to the request.server_name value, this caused the assertion
to fail.

Ideally we'd be able to set the SERVER_NAME in the config for all
platforms, but I was unable to find the correct config value for this.
So in the mean time I'm just accepting both options as valid.