Fix failing annotator login test on OSX
The SERVER_NAME environment variable on OSX defaults to 127.0.0.1 where
on other systems it's localhost. When a user is created the provider
defaults to the request.server_name value, this caused the assertion
to fail.

Secondly, Selenium only returns visible text when using the element.text
property although some drivers (FireFox) seem to return hidden text
also. So now we use dom properties to access the contents of the text
node regardless of styling.