Restore default view to 'Screen'
As a part of the fix for #1231 in commit
https://github.com/hypothesis/h/commit/f66448fc3b2d1e07cbb491db73367b4b69e5f9a6
the default view was change to 'Document' from 'Screen'
but the annotations were not loaded into the sidebar as supposed
for 'Document' view, because the applyView was called before
the annotations were loaded.

Since it'd take effort to correctly implement Document view update
when the store loads new annotations, it is feasible to return
to the original 'Screen' view, which works well, because
views are planned to be refined, maybe redesigned.

Fix #1276